text_to_image/requirements.txt
# Text-to-Image Generator dependencies
flask==2.3.3
torch==2.0.1
diffusers==0.21.4
transformers==4.34.0
accelerate==0.21.0
safetensors==0.3.3
pillow==10.0.0
xformers==0.0.22